refactor: remove legacy advanced pool details
