--- apps/main/src/dex/components/PagePool/PoolDetails/CurrencyReserves/index.tsx
@@ -1,73 +1,0 @@
-import { styled } from 'styled-components'
-import { CurrencyReservesContent } from '@/dex/components/PagePool/PoolDetails/CurrencyReserves/CurrencyReservesContent'
-import { StyledStats } from '@/dex/components/PagePool/PoolDetails/PoolStats/styles'
-import { useNetworkByChain } from '@/dex/entities/networks'
-import { usePoolTokensLinksMapper } from '@/dex/hooks/usePoolTokensLinksMapper'
-import { usePoolTvl } from '@/dex/queries/pool-tvl.query'
-import { useStore } from '@/dex/store/useStore'
-import { ChainId, TokensMapper } from '@/dex/types/main.types'
-import { getChainPoolIdActiveKey } from '@/dex/utils'
-import type { Address } from '@primitives/address.utils'
-import { TooltipIcon as IconTooltip } from '@ui/Tooltip/TooltipIcon'
-import { Chip } from '@ui/Typography'
-import { t } from '@ui-kit/lib/i18n'
-import { copyToClipboard, formatNumber, amount } from '@ui-kit/utils'
-
-type Props = {
-  chainId: ChainId
-  poolId: string
-  tokensMapper: TokensMapper
-}
-
-export const CurrencyReserves = ({ chainId, poolId, tokensMapper }: Props) => {
-  const { data: network } = useNetworkByChain({ chainId })
-  const poolDataMapperCached = useStore(state => state.storeCache.poolsMapper[chainId]?.[poolId])
-  const poolData = useStore(state => state.pools.poolsMapper[chainId]?.[poolId])
-  const currencyReserves = useStore(state => state.pools.currencyReserves[getChainPoolIdActiveKey(chainId, poolId)])
-
-  const poolDataCachedOrApi = poolData ?? poolDataMapperCached
-  const poolTokensLinks = usePoolTokensLinksMapper(chainId, poolDataCachedOrApi)
-
-  const { data: tvl } = usePoolTvl({ chainId, poolId })
-
-  return (
-    <article>
-      <StyledTitle>{t`Currency reserves`}</StyledTitle>
-      {poolDataCachedOrApi?.tokens.map((token, idx) => {
-        const tokenAddress = poolDataCachedOrApi.tokenAddresses[idx]
-        return (
-          <CurrencyReservesContent
-            // eslint-disable-next-line @eslint-react/no-array-index-key -- Existing violation before enabling this rule.
-            key={`${token}-${idx}`}
-            cr={currencyReserves?.tokens.find(t => t.tokenAddress === tokenAddress)}
-            haveSameTokenName={poolDataCachedOrApi.tokensCountBy[token] > 1}
-            network={network}
-            rChainId={chainId}
-            tokensMapper={tokensMapper}
-            token={token}
-            tokenAddress={tokenAddress}
-            tokenLink={poolTokensLinks?.[tokenAddress as Address]}
-            // eslint-disable-next-line @typescript-eslint/no-misused-promises -- Existing violation before enabling this rule.
-            handleCopyClick={copyToClipboard}
-          />
-        )
-      })}
-
-      <StyledStats flex flexJustifyContent="space-between">
-        {t`USD total`}
-        <StyledChip size="md">
-          {formatNumber(amount(tvl), 'usd.amount')}{' '}
-          <IconTooltip placement="bottom-end">{t`USD total balance updates every ~5 minute`}</IconTooltip>
-        </StyledChip>
-      </StyledStats>
-    </article>
-  )
-}
-
-const StyledTitle = styled.h3`
-  margin-bottom: var(--spacing-2);
-`
-
-const StyledChip = styled(Chip)`
-  padding: var(--spacing-2) 0;
-`
