--- apps/main/src/dex/components/PagePool/PoolDetails/CurrencyReserves/types.ts
@@ -1,13 +1,0 @@
-import { ChainId, NetworkConfig, CurrencyReservesToken, TokensMapper } from '@/dex/types/main.types'
-
-export type CurrencyReservesProps = {
-  cr: CurrencyReservesToken | undefined
-  haveSameTokenName: boolean
-  network: NetworkConfig
-  rChainId: ChainId
-  token: string
-  tokenAddress: string
-  tokenLink: string | undefined
-  tokensMapper: TokensMapper
-  handleCopyClick: (address: string) => void
-}
