+++ packages/curve-ui-kit/src/hooks/useFeatureFlags.ts
@@ -47,6 +47,3 @@ export const useNewLlamalendHealth = useBetaChannel
 /** New DEX pool list backed by Prices API v2 */
 export const useDexPoolListV2 = useBetaChannel
 export const isDexPoolListV2Enabled = (releaseChannel: ReleaseChannel) => releaseChannel === ReleaseChannel.Beta
-
-/** Fresh pool page experience */
-export const usePoolFreshup = useStableChannel
